+++ src/Resources/contao/config/config.php
@@ -19,6 +19,7 @@ $GLOBALS['BE_MOD']['content']['fideid'] = array
 	'tables'            => array('tl_fideid', 'tl_fideid_templates', 'tl_fideid_mails'),
 	'icon'              => 'bundles/contaofideid/images/icon.png',
 	'send'              => array('Schachbulle\ContaoFideidBundle\Classes\Mailer', 'send'), 
+	'getMailbuttons'    => array('Schachbulle\ContaoFideidBundle\Classes\Speedmailer', 'send'),
 );
 
 /**