+++ src/Resources/contao/languages/de/default.php
@@ -1,0 +1,3 @@
+<?php
+
+$GLOBALS['TL_LANG']['Fideid']['emailCorrupt'] = 'Die Adresse %s enthält ungültige Zeichen!';