+++ src/Resources/contao/languages/de/tl_fideid.php
@@ -55,8 +55,7 @@ $GLOBALS['TL_LANG']['tl_fideid']['bemerkungen'] = array('Bemerkung', 'Bemerkunge
 $GLOBALS['TL_LANG']['tl_fideid']['intern_legend'] = 'Interne Bemerkungen';
 $GLOBALS['TL_LANG']['tl_fideid']['intern'] = array('Bemerkungen', 'Interne Bemerkungen');
 
-$GLOBALS['TL_LANG']['tl_fideid']['publish_legend'] = 'Bearbeitungsstatus';
-$GLOBALS['TL_LANG']['tl_fideid']['published'] = array('Fertig', 'Der Antrag wurde fertig bearbeitet.');
+$GLOBALS['TL_LANG']['tl_fideid']['speedmail_legend'] = 'Schnelle E-Mail-Kommunikation';
 
 /**
  * Buttons für globale Operationen
@@ -70,7 +69,7 @@ $GLOBALS['TL_LANG']['tl_fideid']['templates'] = array('E-Mail-Templates', 'E-Mai
 
 $GLOBALS['TL_LANG']['tl_fideid']['new'] = array('Neuer Antrag', 'Neuen Antrag anlegen');
 $GLOBALS['TL_LANG']['tl_fideid']['edit'] = array('Antrag bearbeiten', 'Antrag %s bearbeiten');
-$GLOBALS['TL_LANG']['tl_fideid']['emailbox'] = array('Postfach', 'E-Mail-Postfach verwalten');
+$GLOBALS['TL_LANG']['tl_fideid']['emailbox'] = array('Postfach', 'E-Mail-Postfach bearbeiten');
 $GLOBALS['TL_LANG']['tl_fideid']['copy'] = array('Antrag kopieren', 'Antrag %s kopieren');
 $GLOBALS['TL_LANG']['tl_fideid']['delete'] = array('Antrag löschen', 'Antrag %s löschen');
 $GLOBALS['TL_LANG']['tl_fideid']['toggle'] = array('Antrag aktivieren/deaktivieren', 'Antrag %s aktivieren/deaktivieren');
@@ -117,4 +116,8 @@ $GLOBALS['TL_LANG']['tl_fideid']['geschlecht_optionen'] = array
  */
 
 $GLOBALS['TL_LANG']['tl_fideid']['imagickNotInstalled'] = 'PDF-Vorschau nicht möglich. Imagick nicht installiert.';
+
+$GLOBALS['TL_LANG']['tl_fideid']['emailbox_versendet'] = ' (%s alte E-Mails vorhanden)';
+$GLOBALS['TL_LANG']['tl_fideid']['emailbox_offen'] = ' (%s unversendete E-Mails, %s alte E-Mails vorhanden)';
+$GLOBALS['TL_LANG']['tl_fideid']['emailbox_leer'] = ' (Keine E-Mails vorhanden)';