+++ src/Resources/contao/languages/de/tl_fideid_mails.php
@@ -20,13 +20,16 @@ $GLOBALS['TL_LANG']['tl_fideid_mails']['show']= array("Details zur E-Mail %s anz
 
 // Formularfelder
 $GLOBALS['TL_LANG']['tl_fideid_mails']['text_legend'] = 'Vorlage und Text';
-$GLOBALS['TL_LANG']['tl_fideid_mails']['subject']= array('Betreff', 'Betreffzeile der E-Mail');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['default_subject']= array('Betreff', 'Betreffzeile aus dem Template.');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['subject']= array('Betreff (optional)', 'Diese Betreffzeile überschreibt die im Template definierte Betreffzeile. Lassen Sie diese Betreffzeile frei, wenn Sie die voreingestellte Betreffzeile des Templates nutzen wollen.');
 $GLOBALS['TL_LANG']['tl_fideid_mails']['content']= array('Text', 'Textinhalt der E-Mail');
 
 $GLOBALS['TL_LANG']['tl_fideid_mails']['template_legend'] = 'Vorschau';
 $GLOBALS['TL_LANG']['tl_fideid_mails']['template']= array('E-Mail-Template', 'Ein Templatewechsel ändert automatisch die Vorschau');
 $GLOBALS['TL_LANG']['tl_fideid_mails']['signatur']= array('Signatur', 'Standardsignatur verwenden');
 $GLOBALS['TL_LANG']['tl_fideid_mails']['preview']= array('Vorschau', 'Vorschau mit diesem Template');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['preview_subject']= array('Betreff', 'Betreffzeile der E-Mail');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['preview_text']= array('Text', 'Text der E-Mail');
 
 $GLOBALS['TL_LANG']['tl_fideid_mails']['mail_legend'] = 'E-Mail-Optionen';
 $GLOBALS['TL_LANG']['tl_fideid_mails']['insertLizenz'] = array('Lizenzurkunde DIN A4 einfügen', 'PDF der Lizenzurkunde im Format DIN A4 als Anhang mitschicken');